requirements.txt
tensorflow==1.15.0
torch==1.8.0
torchvision==0.9.0
alembic==1.0.5
asn1crypto==0.24.0
blinker==1.4
# cffi==1.11.5
Click==7.0
cryptography==3.2
Flask==1.0.2
Flask-Cors==3.0.10
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
requests
numpy
Pillow
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.2
python-dateutil==2.7.5
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.14
Werkzeug==0.15.3
WTForms==2.2.1